+++ build.config.json
@@ -1,0 +1,15 @@
+{
+	"appId": "com.example.app",
+	"productName": "Svelte Electron",
+	"asar": false,
+	"directories": {
+		"output": "dist"
+	},
+	"files": [
+		"src/electron.cjs",
+		{
+			"from": "build",
+			"to": ""
+		}
+	]
+}
