+++ src/electron.cjs
@@ -1,0 +1,97 @@
+const windowStateManager = require('electron-window-state');
+const contextMenu = require('electron-context-menu');
+const { app, BrowserWindow } = require('electron');
+const serve = require('electron-serve');
+
+try {
+	require('electron-reloader')(module);
+} catch (e) {
+	console.error(e);
+}
+
+const serveURL = serve({ directory: "." });
+const port = process.env.PORT || 3000;
+const dev = !app.isPackaged;
+let mainWindow;
+
+function createWindow() {
+	let windowState = windowStateManager({
+		defaultWidth: 800,
+		defaultHeight: 600,
+	});
+	
+	const mainWindow = new BrowserWindow({
+		backgroundColor: 'whitesmoke',
+		titleBarStyle: 'hidden',
+		autoHideMenuBar: true,
+		trafficLightPosition: {
+			x: 17,
+			y: 32,
+		},
+		minHeight: 450,
+		minWidth: 500,
+		webPreferences: {
+			enableRemoteModule: true,
+			contextIsolation: true,
+			nodeIntegration: true,
+			spellcheck: false,
+			devTools: dev,
+		},
+		x: windowState.x,
+		y: windowState.y,
+		width: windowState.width,
+		height: windowState.height,
+	});
+
+	windowState.manage(mainWindow);
+
+	mainWindow.once('ready-to-show', () => {
+		mainWindow.show();
+		mainWindow.focus();
+	});
+
+	mainWindow.on('close', () => {
+		windowState.saveState(mainWindow);
+	});
+
+	return mainWindow;
+}
+
+contextMenu({
+	showLookUpSelection: false,
+	showSearchWithGoogle: false,
+	showCopyImage: false,
+	prepend: (defaultActions, params, browserWindow) => [
+		{
+			label: 'Make App 💻',
+		},
+	],
+});
+
+function loadVite(port) {
+	mainWindow.loadURL(`http://127.0.0.1:${port}`).catch((e) => {
+		console.log('Error loading URL, retrying', e);
+		setTimeout(() => {
+			loadVite(port);
+		}, 200);
+	});
+}
+
+function createMainWindow() {
+	mainWindow = createWindow();
+	mainWindow.once('close', () => { mainWindow = null });
+
+	dev
+	? loadVite(port)
+	: serveURL(mainWindow)
+}
+
+app.once('ready', createMainWindow);
+app.on('activate', () => {
+	if (!mainWindow) {
+		createMainWindow();
+	}
+});
+app.on('window-all-closed', () => {
+	if (process.platform !== 'darwin') app.quit();
+});
